requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
pydantic[email]
python-dotenv
passlib
python-multipar
